// packages/core/src/gen/google/type/month_pb.ts
// Copyright 2025 Google LLC
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=.js,ts_nocheck=false"
// @generated from file google/type/month.proto (package google.type, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc } from "@bufbuild/protobuf/codegenv2";

/**
 * Describes the file google/type/month.proto.
 */
export const file_google_type_month: GenFile = /*@__PURE__*/
  fileDesc("Chdnb29nbGUvdHlwZS9tb250aC5wcm90bxILZ29vZ2xlLnR5cGUqsAEKBU1vbnRoEhUKEU1PTlRIX1VOU1BFQ0lGSUVEEAASCwoHSkFOVUFSWRABEgwKCEZFQlJVQVJZEAISCQoFTUFSQ0gQAxIJCgVBUFJJTBAEEgcKA01BWRAFEggKBEpVTkUQBhIICgRKVUxZEAcSCgoGQVVHVVNUEAgSDQoJU0VQVEVNQkVSEAkSCwoHT0NUT0JFUhAKEgwKCE5PVkVNQkVSEAsSDAoIREVDRU1CRVIQDEJdCg9jb20uZ29vZ2xlLnR5cGVCCk1vbnRoUHJvdG9QAVo2Z29vZ2xlLmdvbGFuZy5vcmcvZ2VucHJvdG8vZ29vZ2xlYXBpcy90eXBlL21vbnRoO21vbnRoogIDR1RQYgZwcm90bzM");

/**
 * Represents a month in the Gregorian calendar.
 *
 * @generated from enum google.type.Month
 */
export enum Month {
  /**
   * The unspecified month.
   *
   * @generated from enum value: MONTH_UNSPECIFIED = 0;
   */
  MONTH_UNSPECIFIED = 0,

  /**
   * The month of January.
   *
   * @generated from enum value: JANUARY = 1;
   */
  JANUARY = 1,

  /**
   * The month of February.
   *
   * @generated from enum value: FEBRUARY = 2;
   */
  FEBRUARY = 2,

  /**
   * The month of March.
   *
   * @generated from enum value: MARCH = 3;
   */
  MARCH = 3,

  /**
   * The month of April.
   *
   * @generated from enum value: APRIL = 4;
   */
  APRIL = 4,

  /**
   * The month of May.
   *
   * @generated from enum value: MAY = 5;
   */
  MAY = 5,

  /**
   * The month of June.
   *
   * @generated from enum value: JUNE = 6;
   */
  JUNE = 6,

  /**
   * The month of July.
   *
   * @generated from enum value: JULY = 7;
   */
  JULY = 7,

  /**
   * The month of August.
   *
   * @generated from enum value: AUGUST = 8;
   */
  AUGUST = 8,

  /**
   * The month of September.
   *
   * @generated from enum value: SEPTEMBER = 9;
   */
  SEPTEMBER = 9,

  /**
   * The month of October.
   *
   * @generated from enum value: OCTOBER = 10;
   */
  OCTOBER = 10,

  /**
   * The month of November.
   *
   * @generated from enum value: NOVEMBER = 11;
   */
  NOVEMBER = 11,

  /**
   * The month of December.
   *
   * @generated from enum value: DECEMBER = 12;
   */
  DECEMBER = 12,
}

/**
 * Describes the enum google.type.Month.
 */
export const MonthSchema: GenEnum<Month> = /*@__PURE__*/
  enumDesc(file_google_type_month, 0);